requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.9.3
python-decouple==3.6
gunicorn==20.1.0
django-heroku==0.3.1
